Time: 4 ms (97.60%), Space: 6 MB (33.75%) - LeetHub

+++ 9-palindrome-number/9-palindrome-number.cpp
@@ -4,7 +4,7 @@ public:
         
         if(x < 0) return false;
         long long int rev = 0;
-        int copy = x;
+        long long int copy = x;
         
         while(copy) {
             rev = rev*10 + (copy%10);
